WIP expenses service;

+++ app-users/src/main/kotlin/com/github/arhor/aws/microservices/playground/users/data/model/Budget.kt
@@ -1,8 +1,9 @@
 package com.github.arhor.aws.microservices.playground.users.data.model
 
 import org.springframework.data.relational.core.mapping.Column
+import java.math.BigDecimal
 
 data class Budget(
     @Column("limit")
-    val limit: Double,
+    val limit: BigDecimal,
 )
