+++ app-users/src/main/kotlin/com/github/arhor/aws/microservices/playground/users/service/dto/UserResponseDto.kt
@@ -1,7 +1,9 @@
 package com.github.arhor.aws.microservices.playground.users.service.dto
 
+import java.math.BigDecimal
+
 data class UserResponseDto(
     val id: Long,
     val email: String,
-    val budgetLimit: Double,
+    val budgetLimit: BigDecimal,
 )
