+++ app-users/src/main/kotlin/com/github/arhor/aws/microservices/playground/users/service/dto/UserUpdateRequestDto.kt
@@ -1,6 +1,8 @@
 package com.github.arhor.aws.microservices.playground.users.service.dto
 
+import java.math.BigDecimal
+
 data class UserUpdateRequestDto(
     val password: String?,
-    val budgetLimit: Double?,
+    val budgetLimit: BigDecimal?,
 )
